src/Portfolio_master/requirement.txt
asgiref==3.4.1
Django==2.2.11
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.3.1
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
